+++ src/modules/mavlink/mavlink_messages.cpp
@@ -75,7 +75,6 @@
 #include <uORB/topics/input_rc.h>
 #include <uORB/topics/manual_control_setpoint.h>
 #include <uORB/topics/mavlink_log.h>
-#include <uORB/topics/mount_orientation.h>
 #include <uORB/topics/optical_flow.h>
 #include <uORB/topics/position_controller_status.h>
 #include <uORB/topics/position_setpoint_triplet.h>
@@ -118,6 +117,7 @@ using matrix::wrap_2pi;
 #include "streams/GPS_STATUS.hpp"
 #include "streams/HIGH_LATENCY2.hpp"
 #include "streams/HIL_STATE_QUATERNION.hpp"
+#include "streams/MOUNT_ORIENTATION.hpp"
 #include "streams/OBSTACLE_DISTANCE.hpp"
 #include "streams/ORBIT_EXECUTION_STATUS.hpp"
 #include "streams/PING.hpp"
@@ -4659,75 +4659,6 @@ protected:
 	}
 };
 
-class MavlinkStreamMountOrientation : public MavlinkStream
-{
-public:
-	const char *get_name() const override
-	{
-		return MavlinkStreamMountOrientation::get_name_static();
-	}
-
-	static constexpr const char *get_name_static()
-	{
-		return "MOUNT_ORIENTATION";
-	}
-
-	static constexpr uint16_t get_id_static()
-	{
-		return MAVLINK_MSG_ID_MOUNT_ORIENTATION;
-	}
-
-	uint16_t get_id() override
-	{
-		return get_id_static();
-	}
-
-	static MavlinkStream *new_instance(Mavlink *mavlink)
-	{
-		return new MavlinkStreamMountOrientation(mavlink);
-	}
-
-	unsigned get_size() override
-	{
-		return _mount_orientation_sub.advertised() ? MAVLINK_MSG_ID_MOUNT_ORIENTATION_LEN + MAVLINK_NUM_NON_PAYLOAD_BYTES : 0;
-	}
-
-private:
-	uORB::Subscription _mount_orientation_sub{ORB_ID(mount_orientation)};
-	uORB::Subscription _lpos_sub{ORB_ID(vehicle_local_position)};
-
-	/* do not allow top copying this class */
-	MavlinkStreamMountOrientation(MavlinkStreamMountOrientation &) = delete;
-	MavlinkStreamMountOrientation &operator = (const MavlinkStreamMountOrientation &) = delete;
-
-protected:
-	explicit MavlinkStreamMountOrientation(Mavlink *mavlink) : MavlinkStream(mavlink)
-	{}
-
-	bool send() override
-	{
-		mount_orientation_s mount_orientation;
-
-		if (_mount_orientation_sub.update(&mount_orientation)) {
-			mavlink_mount_orientation_t msg{};
-
-			msg.roll = math::degrees(mount_orientation.attitude_euler_angle[0]);
-			msg.pitch = math::degrees(mount_orientation.attitude_euler_angle[1]);
-			msg.yaw = math::degrees(mount_orientation.attitude_euler_angle[2]);
-
-			vehicle_local_position_s lpos{};
-			_lpos_sub.copy(&lpos);
-			msg.yaw_absolute = math::degrees(matrix::wrap_2pi(lpos.heading + mount_orientation.attitude_euler_angle[2]));
-
-			mavlink_msg_mount_orientation_send_struct(_mavlink->get_channel(), &msg);
-
-			return true;
-		}
-
-		return false;
-	}
-};
-
 static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamHeartbeat>(),
 	create_stream_list_item<MavlinkStreamStatustext>(),
@@ -4793,7 +4724,9 @@ static const StreamListItem streams_list[] = {
 	create_stream_list_item<MavlinkStreamUTMGlobalPosition>(),
 	create_stream_list_item<MavlinkStreamCollision>(),
 	create_stream_list_item<MavlinkStreamWind>(),
+#if defined(MOUNT_ORIENTATION_HPP)
 	create_stream_list_item<MavlinkStreamMountOrientation>(),
+#endif // MOUNT_ORIENTATION_HPP
 #if defined(HIGH_LATENCY2_HPP)
 	create_stream_list_item<MavlinkStreamHighLatency2>(),
 #endif // HIGH_LATENCY2_HPP
